+++ grease_macro/src/lib.rs
@@ -685,7 +685,7 @@ fn do_grease_trait_impl(
 
     let func_sets = funcs.iter().map(|(name, _, _, _, _)| {
         quote_spanned! { name.span() =>
-            #name: ::grease::closure::FnPtr::new(#name #type_generics)
+            #name: unsafe { ::grease::closure::FnPtr::new(#name #type_generics) }
         }
     });